// src/i & a/identity/domain/enum/identityCapabilities.enum.ts
enum IdentityCapabilities {
	STAFF_VIEW = "staff.view",
	STAFF_LIST = "staff.list",
	STAFF_CREATE = "staff.create",
	STAFF_UPDATE = "staff.update",
	STAFF_ACTIVATE = "staff.activate",
	STAFF_DEACTIVATE = "staff.deactivate",
	STAFF_PENDING_ACTIVATION_LIST = "staff.pending_activation.list",
	CLASSIFICATION_ASSIGN = "classification.assign",
	CLASSIFICATION_UPDATE = "classification.update",
	CLASSIFICATION_VIEW = "classification.view",
	ORGANIZATION_MANAGE = "organization.manage",
}

export { IdentityCapabilities };